cleanix/pyproject.toml
[tool.poetry]
name = "cleanix"
version = "0.1.0"
description = ""
authors = ["Linyshka <lina.radyukevich98@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
aiosql = "^9.0"
pydantic-settings = "^2.1.0"
pydantic = "^2.5.2"
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
